@@ -1,20 +1,14 @@
-repositories {
-    maven { url 'https://jitpack.io' }
-}
-
-apply plugin: 'com.google.gms.google-services'
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '26.0.3'
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.ramotion.showroom"
         minSdkVersion 19
-        targetSdkVersion 26
-        versionCode 3
-        versionName "1.1.2"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        targetSdkVersion 28
+        versionCode 4
+        versionName "1.1.3"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -35,36 +29,37 @@ android {
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
     // Google Analytics
-    compile 'com.google.android.gms:play-services-analytics:11.8.0'
+    implementation 'com.google.android.gms:play-services-analytics:16.0.8'
     // Support library
-    compile 'com.android.support:appcompat-v7:26.1.0'
-    compile 'com.android.support:design:26.1.0'
-    compile 'com.android.support:support-v4:26.1.0'
-    compile 'com.android.support:cardview-v7:26.1.0'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'com.google.android.material:material:1.1.0-alpha05'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     // Controls
-    compile 'com.ramotion.expandingcollection:expanding-collection:0.9.0'
-    compile 'com.ramotion.foldingcell:folding-cell:1.2.1'
-    compile 'com.ramotion.paperonboarding:paper-onboarding:1.1.1'
-    compile 'com.ramotion.cardslider:card-slider:0.1.0'
-    compile 'com.ramotion.garlandview:garland-view:0.2.1'
-    compile 'com.ramotion.circlemenu:circle-menu:0.1.0'
+    implementation 'com.ramotion.expandingcollection:expanding-collection:0.9.2'
+    implementation 'com.ramotion.foldingcell:folding-cell:1.2.3'
+    implementation 'com.ramotion.paperonboarding:paper-onboarding:1.1.3'
+    implementation 'com.ramotion.cardslider:card-slider:0.3.1'
+    implementation 'com.ramotion.garlandview:garland-view:0.3.3'
+    implementation 'com.ramotion.circlemenu:circle-menu:0.3.2'
     // For GarlandView example
-    compile 'io.reactivex.rxjava2:rxandroid:2.0.1'
-    compile 'io.reactivex.rxjava2:rxjava:2.1.2'
-    compile 'com.github.blocoio:faker:1.2.5'
-    compile 'com.github.bumptech.glide:glide:3.8.0'
-    compile 'jp.wasabeef:glide-transformations:2.0.2'
-    compile 'org.greenrobot:eventbus:3.0.0'
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.0'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.0'
+    implementation 'com.github.blocoio:faker:1.2.5'
+    implementation 'com.github.bumptech.glide:glide:4.8.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.8.0'
+    implementation 'jp.wasabeef:glide-transformations:2.0.2'
+    implementation 'org.greenrobot:eventbus:3.0.0'
     // Testing
-    testCompile 'junit:junit:4.12'
     // Leak Canary
-    debugCompile 'com.squareup.leakcanary:leakcanary-android:1.5.4'
-    releaseCompile 'com.squareup.leakcanary:leakcanary-android-no-op:1.5.4'
-    
+    debugImplementation 'com.squareup.leakcanary:leakcanary-android:1.5.4'
+    releaseImplementation 'com.squareup.leakcanary:leakcanary-android-no-op:1.5.4'
+
+    testImplementation 'junit:junit:4.12'
+    androidTestImplementation 'androidx.test:runner:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
 }
+
+apply plugin: 'com.google.gms.google-services'
