@@ -1,7 +1,6 @@
 package com.ramotion.showroom.examples.garlandview;
 
-
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.lang.ref.WeakReference;
 
