@@ -2,7 +2,6 @@ package com.ramotion.showroom;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v7.widget.CardView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -12,9 +11,12 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.cardview.widget.CardView;
+import androidx.viewpager.widget.PagerAdapter;
+
 import java.util.ArrayList;
 
-public class SliderAdapter extends android.support.v4.view.PagerAdapter {
+public class SliderAdapter extends PagerAdapter {
     private Context context;
 
     private ArrayList<SlideCardEntity> dataset = new ArrayList<>();
