@@ -1,10 +1,11 @@
 package com.ramotion.showroom.examples.cardslider.cards;
 
 
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.ramotion.showroom.R;
 
