@@ -2,13 +2,14 @@ package com.ramotion.showroom.examples.expandingcollection;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.ramotion.expandingcollection.ECCardContentListItemAdapter;
 import com.ramotion.showroom.R;
