@@ -1,9 +1,10 @@
 package com.ramotion.showroom.examples.garlandview.details;
 
-import android.databinding.DataBindingUtil;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
+
+import androidx.databinding.DataBindingUtil;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.ramotion.showroom.R;
 import com.ramotion.showroom.databinding.GvDetailsItemBinding;
